% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankgm.R
\name{dnbmix}
\alias{dnbmix}
\title{Negative binomial mixture probability density}
\usage{
dnbmix(theta, y, dsum = TRUE, weights = TRUE)
}
\arguments{
\item{theta}{Numeric vector of parameter values \eqn{w_1, n_1, p_1, w_2, n_2, p_2, ..., p_{k-1}, n_k, p_k}}

\item{y}{Numeric vector of gene counts from a single cell, including zeros}

\item{dsum}{Logical, whether to sum the probabilities across the distributions, default is TRUE}

\item{weights}{Logical, whether to weight the probabilities according to the parameters \eqn{w_1, w_2, ...,
w_{k-1}}, default is TRUE}
}
\value{
Returns a matrix of probability densities for the given gene counts, possibly weighted and/or summed
  across the mixture of distributions.
}
\description{
Probability density for the mixture of \eqn{k} negative binomial distributions: \eqn{w_1 NB(n_1, p_1) + w_2
NB(n_2, p_2) + ... + (1 - \Sigma w) NB(n_k, p_k)}
}
\keyword{internal}
