% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankgm.R
\name{grLL}
\alias{grLL}
\title{Log-likelihood gradient}
\usage{
grLL(theta, y, clip.scale = 20000)
}
\arguments{
\item{theta}{Numeric vector of parameter values, see `dnbmix` for details}

\item{y}{Numeric vector of gene counts from a single cell, including zeros}

\item{clip.scale}{Maximum allowed magnitude of the gradient, relative to the current parameter value}
}
\value{
Returns clipped values of the log-likelihood gradient: \eqn{\nabla \ell = \frac{\partial \ell}
  {\partial w_1}, \frac{\partial \ell}{\partial n_1}, \frac{\partial \ell}{\partial p_1}, \frac{\partial
  \ell}{\partial w_2}, \frac{\partial \ell}{\partial n_2}, \frac{\partial \ell}{\partial p_2}, ...,
  \frac{\partial \ell}{\partial p_{k-1}}, \frac{\partial \ell}{\partial n_k}, \frac{\partial \ell}
  {\partial p_k}}
}
\description{
Computes \eqn{\frac{\partial \ell}{\partial \theta}} for the gene counts \eqn{y} and parameters \eqn{\theta}
}
\keyword{internal}
