% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankgm.R
\name{inferRanks}
\alias{inferRanks}
\title{Infer gene expression-module rankings}
\usage{
inferRanks(y.nz, n.genes.total, thresh.p = 0.05, max.n = 10, seed = 1234)
}
\arguments{
\item{y.nz}{List of all non-zero gene counts from a given cell}

\item{n.genes.total}{Total number of genes in the given cell, including zeros}

\item{thresh.p}{The p-value threshold for adding new distributions to the mixture model, see `lkrTest` for details}

\item{max.n}{The maximum number of distributions to include in the final mixture model}

\item{seed}{Fixed seed for reproducibility, default is 1234}
}
\value{
Returns a single numeric vector containing the single-cell-specific gene rankings for the non-zero gene
  counts, followed by a binary indicator of the success (1) or failure (0) of the model-fitting procedure. The
  ranks are returned as all zeros if the fitting procedure is unsuccessful.
}
\description{
Fits a series of negative binomial mixture models to the gene counts from a single cell to first infer and rank the
underlying gene-expression modules, then rank-order genes according to their most likely module.
}
\keyword{internal}
