% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankgm.R
\name{lkrTest}
\alias{lkrTest}
\title{Likelihood-ratio test}
\usage{
lkrTest(f1, f2)
}
\arguments{
\item{f1}{Fitted model, see `optim` in `stats` for more details}

\item{f2}{See f1}
}
\value{
Returns the probability of rejecting the null hypothesis from a one-tailed chi-squared distribution
}
\description{
Uses a chi-square test to compute the probability of rejecting the null hypothesis that the simpler ('f1') and more
complex ('f2') models both fit the data equally well, given the ratio of the models' likelihoods.
}
\keyword{internal}
