% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankgm.R
\name{negLL}
\alias{negLL}
\title{Negative log-likelihood}
\usage{
negLL(theta, y)
}
\arguments{
\item{theta}{Numeric vector of parameter values, see `dnbmix` for details}

\item{y}{Numeric vector of gene counts from a single cell, including zeros}
}
\value{
Returns the negative log-likelihood value for \eqn{\theta|y}
}
\description{
Computes \eqn{-\ell(\theta|y)} for the gene counts \eqn{y} and parameters \eqn{\theta}
}
\keyword{internal}
